+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/fragment/review/RestaurantReviewInfoFragmentTest.kt
@@ -1,14 +1,18 @@
-package com.github.sdp.ratemyepfl.activity
+package com.github.sdp.ratemyepfl.fragment.review
 
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.withText
 import com.github.sdp.ratemyepfl.R
+import com.github.sdp.ratemyepfl.activity.ReviewActivity
+import com.github.sdp.ratemyepfl.database.FakeCourseRepository
 import com.github.sdp.ratemyepfl.database.FakeRestaurantRepository
+import com.github.sdp.ratemyepfl.database.FakeReviewsRepository
+import com.github.sdp.ratemyepfl.utils.CustomViewActions
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.After
@@ -17,7 +21,7 @@ import org.junit.Rule
 import org.junit.Test
 
 @HiltAndroidTest
-class RestaurantReviewActivityTest {
+class RestaurantReviewInfoFragmentTest {
     lateinit var scenario: ActivityScenario<ReviewActivity>
 
     @get:Rule(order = 0)
@@ -36,13 +40,27 @@ class RestaurantReviewActivityTest {
         scenario.close()
     }
 
-
     @Test
-    fun isIdVisibleOnActivityLaunch() {
+    fun allInformationCorrectlyDisplayed() {
         val fakeRestaurant = FakeRestaurantRepository.DEFAULT_RESTAURANT
+        val fakeReviewList = FakeReviewsRepository.fakeList
+        val numReviewText = "(${fakeReviewList.size} reviews)"
         onView(withId(R.id.restaurantIdInfo))
-            .check(matches(ViewMatchers.withText(fakeRestaurant.toString())))
+            .check(matches(withText(fakeRestaurant.id)))
+        onView(withId(R.id.restaurantNumReview)).check(matches(withText(numReviewText)))
     }
 
+    @Test
+    fun noReviewDisplayed(){
+        FakeReviewsRepository.reviewList = listOf()
+
+        // Refresh
+        onView(withId(R.id.reviewNavigationView)).perform(CustomViewActions.navigateTo(R.id.addReviewFragment))
+        onView(withId(R.id.reviewNavigationView)).perform(CustomViewActions.navigateTo(R.id.restaurantReviewInfoFragment))
+
+        val numReviewText = "(No review submitted)"
+        onView(withId(R.id.restaurantNumReview)).check(matches(withText(numReviewText)))
+        FakeReviewsRepository.reviewList = FakeReviewsRepository.fakeList
+    }
 
 }
